refactor(config): Make config a class and each type of config (e.g., toml, ini) a child class
* Use new config structure to fix the previously fixed Issue #97
* Rename the config inside a config as settings
* Use config class as argument for cz and commands instead of dict config
* Remove settings from default
