+++ commitizen/cz/base.py
@@ -22,8 +22,8 @@ class BaseCommitizen(metaclass=ABCMeta):
 
     def __init__(self, config: dict):
         self.config = config
-        if not self.config.get("style"):
-            self.config["style"] = BaseCommitizen.default_style_config
+        if not self.config.settings.get("style"):
+            self.config._settings["style"] = BaseCommitizen.default_style_config
 
     @abstractmethod
     def questions(self) -> list:
@@ -36,7 +36,10 @@ class BaseCommitizen(metaclass=ABCMeta):
     @property
     def style(self):
         return merge_styles(
-            [Style(BaseCommitizen.default_style_config), Style(self.config["style"])]
+            [
+                Style(BaseCommitizen.default_style_config),
+                Style(self.config._settings["style"]),
+            ]
         )
 
     def example(self) -> str:
