+++ commitizen/defaults.py
@@ -2,15 +2,6 @@ name: str = "cz_conventional_commits"
 # TODO: .cz, setup.cfg, .cz.cfg should be removed in 2.0
 config_files: list = ["pyproject.toml", ".cz.toml", ".cz", "setup.cfg", ".cz.cfg"]
 
-# Available settings
-settings: dict = {
-    "name": name,
-    "version": None,
-    "files": [],
-    "tag_format": None,  # example v$version
-    "bump_message": None,  # bumped v$current_version to $new_version
-}
-
 MAJOR = "MAJOR"
 MINOR = "MINOR"
 PATCH = "PATCH"
